requirements.txt
anyio==3.7.1
azure-cognitiveservices-speech==1.43.0
azure-core==1.34.0
azure-cosmos==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.104.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
numpy==2.2.5
openai==1.3.0
openpyxl==3.1.5
pandas==2.2.3
pydantic==1.10.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0